Separate s3urlcache into it’s own file.
